Table S2. Module memberships of consumer and resource trait modalities in all four habitat trait webs.
Table S2. Module memberships of consumer and resource trait modalities in all four habitat trait webs. ‘TM_#’ is the index of the module member. ‘Count_Modalities’ gives the number of consumer or resource trait modalities involved in a module. ‘Size’ is the product of the number of consumer and resource trait modalities included within the same module. It indicates how many consumer-resource trait associations were possible inside a module, and gives an estimate of the size of a module, for comparability. N.B. Modalities were re-arranged for comparability and their index does not necessarily correspond to the order they appeared on the graph.